OGM-389 Extracting shared test model classes from into separate package
