+++ core/src/test/java/org/hibernate/ogm/test/options/mapping/AnnotationBasedOptionsTest.java
@@ -29,7 +29,8 @@ import org.hibernate.ogm.test.options.examples.EmbedExampleOption;
 import org.hibernate.ogm.test.options.examples.NameExampleOption;
 import org.hibernate.ogm.test.options.examples.annotations.EmbedExample;
 import org.hibernate.ogm.test.options.examples.annotations.NameExample;
-import org.hibernate.ogm.test.options.mapping.SampleOptionModel.SampleGlobalContext;
+import org.hibernate.ogm.test.options.mapping.model.SampleOptionModel;
+import org.hibernate.ogm.test.options.mapping.model.SampleOptionModel.SampleGlobalContext;
 import org.junit.Before;
 import org.junit.Test;
