+++ core/src/test/java/org/hibernate/ogm/test/options/mapping/JPAOptionIntegrationTest.java
@@ -34,8 +34,8 @@ import org.hibernate.ogm.options.spi.OptionsService.OptionsServiceContext;
 import org.hibernate.ogm.test.options.examples.EmbedExampleOption;
 import org.hibernate.ogm.test.options.examples.ForceExampleOption;
 import org.hibernate.ogm.test.options.examples.NameExampleOption;
-import org.hibernate.ogm.test.options.mapping.OptionIntegrationTest.Microwave;
-import org.hibernate.ogm.test.options.mapping.OptionIntegrationTest.Refrigerator;
+import org.hibernate.ogm.test.options.mapping.model.Microwave;
+import org.hibernate.ogm.test.options.mapping.model.Refrigerator;
 import org.hibernate.ogm.test.utils.PackagingRule;
 import org.junit.Rule;
 import org.junit.Test;
