+++ core/src/test/java/org/hibernate/ogm/test/options/mapping/model/Microwave.java
@@ -20,24 +20,10 @@
  */
 package org.hibernate.ogm.test.options.mapping.model;
 
-import java.lang.annotation.ElementType;
-
-import org.hibernate.ogm.cfg.Configurable;
-import org.hibernate.ogm.cfg.spi.OptionConfigurer;
-
 /**
  * @author Gunnar Morling
  */
-public class SampleOptionConfigurer extends OptionConfigurer {
+public class Microwave {
 
-	@Override
-	public void configure(Configurable configurable) {
-		configurable.configureOptionsFor( SampleNoSqlDatastore.class )
-			.entity( Refrigerator.class )
-				.force( true )
-				.property( "temperature", ElementType.FIELD )
-					.embed( "Embedded" )
-			.entity( Microwave.class )
-				.name( "test" );
-	}
+	public int power;
 }
