+++ core/src/test/java/org/hibernate/ogm/test/options/mapping/model/SampleDatastoreProvider.java
@@ -20,24 +20,22 @@
  */
 package org.hibernate.ogm.test.options.mapping.model;
 
-import java.lang.annotation.ElementType;
-
-import org.hibernate.ogm.cfg.Configurable;
-import org.hibernate.ogm.cfg.spi.OptionConfigurer;
+import org.hibernate.ogm.datastore.spi.DatastoreProvider;
+import org.hibernate.ogm.dialect.GridDialect;
+import org.hibernate.ogm.service.impl.QueryParserService;
 
 /**
  * @author Gunnar Morling
  */
-public class SampleOptionConfigurer extends OptionConfigurer {
+public class SampleDatastoreProvider implements DatastoreProvider {
 
 	@Override
-	public void configure(Configurable configurable) {
-		configurable.configureOptionsFor( SampleNoSqlDatastore.class )
-			.entity( Refrigerator.class )
-				.force( true )
-				.property( "temperature", ElementType.FIELD )
-					.embed( "Embedded" )
-			.entity( Microwave.class )
-				.name( "test" );
+	public Class<? extends GridDialect> getDefaultDialect() {
+		throw new UnsupportedOperationException();
+	}
+
+	@Override
+	public Class<? extends QueryParserService> getDefaultQueryParserServiceType() {
+		throw new UnsupportedOperationException();
 	}
 }
